Fix: Migrate slider to its own component
